Fix Issue where saving same option value errs. (#1433)
Closes issue #1390

There were quite a few places where UPDATES could fail silently because we weren't checking target rows where actually found where we expect them to be. In order to address this problem clientFoundRows was set in the sql driver configuration and checks for UPDATES were added to determine if matched rows were found where we expect them to be.

+++ server/datastore/mysql/email_changes.go
@@ -63,10 +63,19 @@ func (ds *Datastore) ConfirmPendingEmailChange(id uint, token string) (newEmail 
       email = ?
     WHERE id = ?
   `
-	_, err = tx.Exec(query, changeRecord.NewEmail, changeRecord.UserID)
+	results, err := tx.Exec(query, changeRecord.NewEmail, changeRecord.UserID)
 	if err != nil {
 		return "", errors.Wrap(err, "updating user's email")
 	}
+
+	rowsAffected, err := results.RowsAffected()
+	if err != nil {
+		return "", errors.Wrap(err, "fetching affected rows updating user's email")
+	}
+	if rowsAffected == 0 {
+		return "", notFound("User").WithID(changeRecord.UserID)
+	}
+
 	_, err = tx.Exec("DELETE FROM email_changes WHERE id = ?", changeRecord.ID)
 	if err != nil {
 		return "", errors.Wrap(err, "deleting email change")
